remove writing output to file
Spelling errors will only be handled by humans, thus it is not
neccessary to write the output to anything but stdout.
Furthermore, this tool is mainly meant to be used in CI environments, in
which output files will be cleaned by the CI anyways, or will be "hard"
to inspect.

If this feature is ever needed again, reverting this commit should do
the trick.